Delete tag now conditionally fetches

+++ src/hooks/useHandleDeleteTag.ts
@@ -4,27 +4,30 @@ import { useAppDispatch } from "./useAppDispatch"
 import { deleteTag } from "features/tasksSlice"
 import { setSnackbar, setSnackbarError } from "features/appSlice"
 
-export default function useHandleDeleteTag(){ 
+export default function useHandleDeleteTag() {
     const [deleteTagMut, deleteTagResult] = useDeleteTagMutation()
-    const userid = useAppSelector(state => state.auth.userid)   
+    const userid = useAppSelector(state => state.auth.userid)
+    const hasSession = useAppSelector(state => state.auth.hasSession)
     const dispatch = useAppDispatch()
-    async function deleteTagHandler(tagId: number){ 
-        try{
-            const response = deleteTagMut(
-                {
-                  userId: userid!, 
-                  tagId: tagId
-                }
-              ).unwrap()
-            console.log(response)
-            dispatch(deleteTag({id: tagId}))
+    async function deleteTagHandler(tagId: number) {
+        try {
+            if (hasSession && (userid !== null)) {
+                const response = deleteTagMut(
+                    {
+                        userId: userid!,
+                        tagId: tagId
+                    }
+                ).unwrap()
+                console.log(response)
+            }
+            dispatch(deleteTag({ id: tagId }))
             dispatch(setSnackbar("Tag deleted!"))
         }
-        catch(error){ 
+        catch (error) {
             console.log(error)
             dispatch(setSnackbarError("Error deleting tag"))
         }
-    
+
     }
 
     return [deleteTagHandler]
